@@ -28,10 +28,10 @@
         }
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.1",
         "swayok/utils": "^1.3",
         "symfony/http-foundation": ">=3.0",
-        "crazycodr/standard-exceptions": "*"
+        "nesbot/carbon": "^2.0.0"
     },
     "require-dev": {
         "phpunit/phpunit": "8.*"
